Seperate memory state from calldata state

+++ src/test/Events.sol
@@ -2,12 +2,10 @@
 pragma solidity ^0.8.11;
 
 contract Events {
-
     event LogUint(uint256 value);
-    event Paid(uint value);
+    event Paid(uint256 value);
 
     function logUint(uint256 value) public {
         emit LogUint(value);
     }
-
 }
