+++ src/test/Math.sol
@@ -2,10 +2,7 @@
 pragma solidity ^0.8.11;
 
 contract Math {
-    function add(
-        uint256 a,
-        uint256 b
-    ) public pure returns (uint) {
+    function add(uint256 a, uint256 b) public pure returns (uint256) {
         return a + b;
     }
 }
